refactor(pack): refactor wifi
