+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/wifi/wifi.html
